pyproject.toml
[project]
name = "cairn"
version = "0.0.1"
requires-python = ">=3.11"
dependencies = [
    "msgspec>=0.18",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-timeout>=2.2",
]

[project.scripts]
cairn = "cairn.cli:main"
cairnd = "cairn.daemon:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/cairn"]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-ra --timeout=30"
